[harbour-barcode] Fixed stack overflow caused by single CR chars
Basically if the text was containing single CR characters (left after
replacing CRLF with LF) then what we assign to the 'text' property of
TextArea was never becoming equal to the actual value of that property
(because TextEdit appears to be internally replacing single CRs with
LFs too), resulting in infinite recursion and inevitable stack overflow.

Therefore we need to mirror TextEdit behavior and replace single CRs as
well. The whole approach looks a bit fragile (what if TextEdit behavior
changes?) but it seems to be the only way to implement a selectable
read-only text component.